Possible incorrect LCSC part number ( need professional eyes)
the C51386 is a TPS2041BDBVR(logic low) where as the Part number listed -  TPS2051BDBVR is logic high. Not sure were else that should be changed if that is the case. The JLC plugin seems to be keep grabbing the C51386 until i manually selected the part. Not sure what the Plug in references.